Paritally support slurm sbatch command.
Support running a sequence of jobs with sbatch command.
sbatch job saves stdout/stderr to a log file rather than printing it,
So this identifies the logfile location and runs `tail -f` for it.
NOTE: It's quite hard to run `tail -f` on all the logfiles in the proper
order, thus only the first job's stdout/stderr shows up.